internship_projects/Phase-1/Ingestion_Pipeline/requirements.txt
PyPDF2
python-docx
pydantic